requirement.txt
mysql-connector-python==8.0.32
scikit-learn==1.2.2
pandas==1.5.3

